Fixing bugs
1. "Link to Parent Column" partially working. (very inconsistent)
It is working if we create new column/list but not working for existed
columns and that is also inconsistent. Some times disabled and some
times even though the check box is enabled it not showing relevant
column in parent column.

2. "Sort by View" option is not working for simple column. I am unable
to test parent child relation case because of issue 1.

3. "New Element" not auto populating the new entry to the drop down.

4. Page refresh for multiple value enabled lookup column.

5. Unable to create lookup column for the same list. (Example, I am in
Tasks list and I want to choose related task from already existed tasks)

6. Dull UI. With the current UI we can not use this in production.